backend/requirements.txt
flask
flask-cors
gitpython
transformers
sentence-transformers
faiss-cpu
torch
requests